+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.0.1
 
-Initial module plumbing.
+### Added
+
+Initial module version while in development, with the following features:
+
+* Customize chat visibility of initiative rolls, with separate toggles for players and NPCs
+* Show "unrolled" initiative separately from "unknown" initiative in the combat tracker
+* Allow keeping NPC initiative ambiguous, and don't sort them properly in the player-visible table until they've taken a turn